move files under one folder
