+++ src/hamiltonian/diagonalize_fp.hpp
