+++ src/hamiltonian/residuals.hpp
